3.x Clean up traits and remove methods from Date
Inline most of the traits as they only have one usage each. The
FormattingTrait was moved up a directory as having a directory with only
one file feels a bit silly.

I've also marked several classes as internal because I don't think that
userland code should be relying on these classes and logic.
